Add patch method and validation for objectIds

+++ api/crud/src/routes/dish.ts
@@ -38,20 +38,20 @@ router.post("/", async (req, res, next) => {
     const parseResult = await DishSchema.strict().safeParseAsync(req.body);
     if (!parseResult.success) {
       res.status(400).json(parseResult.error.issues);
-    } else {
-      const now = new Date().toJSON();
-      const newItem: Dish & DbMeta = {
-        ...parseResult.data,
-        createdOn: now,
-        lastUpdatedOn: now,
-      };
-      const insertResult = await dishes.insertOne({ ...newItem });
-      const returnItem: Dish & DbMeta & DbId = {
-        ...newItem,
-        id: insertResult.insertedId.toJSON(),
-      };
-      res.status(201).json(returnItem);
+      return;
     }
+    const now = new Date().toJSON();
+    const newItem: Dish & DbMeta = {
+      ...parseResult.data,
+      createdOn: now,
+      lastUpdatedOn: now,
+    };
+    const insertResult = await dishes.insertOne({ ...newItem });
+    const returnItem: Dish & DbMeta & DbId = {
+      ...newItem,
+      id: insertResult.insertedId.toJSON(),
+    };
+    res.status(201).json(returnItem);
   } catch (e) {
     next(e);
   }
@@ -59,39 +59,79 @@ router.post("/", async (req, res, next) => {
 
 router.put("/:objectId", async (req, res, next) => {
   try {
+    if (!ObjectId.isValid(req.params.objectId)) {
+      res
+        .status(400)
+        .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
+      return;
+    }
     const id = new ObjectId(req.params.objectId);
     const parseResult = await DishSchema.strict().safeParseAsync(req.body);
     if (!parseResult.success) {
       res.status(400).json(parseResult.error.issues);
-    } else {
-      const now = new Date().toJSON();
-      const findItem = await dishes.findOne({ _id: id });
-      if (findItem == null) {
-        const newItem: Dish & DbMeta = {
-          ...parseResult.data,
-          createdOn: now,
-          lastUpdatedOn: now,
-        };
-        const insertResult = await dishes.insertOne({ ...newItem, _id: id });
-        const returnItem: Dish & DbMeta & DbId = {
-          ...newItem,
-          id: insertResult.insertedId.toJSON(),
-        };
-        res.status(201).json(returnItem);
-      } else {
-        const updatedItem: Dish & DbMeta = {
-          ...parseResult.data,
-          createdOn: findItem.createdOn,
-          lastUpdatedOn: now,
-        };
-        await dishes.updateOne({ _id: id }, { $set: updatedItem });
-        const returnItem: Dish & DbMeta & DbId = {
-          ...updatedItem,
-          id: findItem._id.toJSON(),
-        };
-        res.status(200).json(returnItem);
-      }
+      return;
     }
+    const now = new Date().toJSON();
+    const findItem = await dishes.findOne({ _id: id });
+    if (findItem === null) {
+      const newItem: Dish & DbMeta = {
+        ...parseResult.data,
+        createdOn: now,
+        lastUpdatedOn: now,
+      };
+      const insertResult = await dishes.insertOne({ ...newItem, _id: id });
+      const returnItem: Dish & DbMeta & DbId = {
+        ...newItem,
+        id: insertResult.insertedId.toJSON(),
+      };
+      res.status(201).json(returnItem);
+    } else {
+      const updatedItem: Dish & DbMeta = {
+        ...parseResult.data,
+        createdOn: findItem.createdOn,
+        lastUpdatedOn: now,
+      };
+      await dishes.updateOne({ _id: id }, { $set: updatedItem });
+      const returnItem: Dish & DbMeta & DbId = {
+        ...updatedItem,
+        id: findItem._id.toJSON(),
+      };
+      res.status(200).json(returnItem);
+    }
+  } catch (e) {
+    next(e);
+  }
+});
+
+router.patch("/:objectId", async (req, res, next) => {
+  try {
+    if (!ObjectId.isValid(req.params.objectId)) {
+      res
+        .status(400)
+        .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
+      return;
+    }
+    const id = new ObjectId(req.params.objectId);
+    const parseResult = await DishSchema.partial()
+      .strict()
+      .safeParseAsync(req.body);
+    if (!parseResult.success) {
+      res.status(400).json(parseResult.error.issues);
+      return;
+    }
+    const findItem = await dishes.findOne({ _id: id });
+    if (findItem === null) {
+      res.status(404).end();
+      return;
+    }
+    const now = new Date().toJSON();
+    const updatedItem: Partial<Dish> & DbMeta = {
+      ...parseResult.data,
+      createdOn: findItem.createdOn,
+      lastUpdatedOn: now,
+    };
+    await dishes.updateOne({ _id: id }, { $set: updatedItem });
+    res.status(204).end();
   } catch (e) {
     next(e);
   }
@@ -99,6 +139,12 @@ router.put("/:objectId", async (req, res, next) => {
 
 router.delete("/:objectId", async (req, res, next) => {
   try {
+    if (!ObjectId.isValid(req.params.objectId)) {
+      res
+        .status(400)
+        .json({ code: "invalid_id", path: "objectId", message: "Invlid ID." });
+      return;
+    }
     const id = new ObjectId(req.params.objectId);
     let deleteResult = await dishes.deleteOne({ _id: id }, {});
     if (deleteResult.deletedCount === 1) {
